+++ src/main/java/com/java_school/cinemabot/model/Film.java
@@ -5,6 +5,7 @@ import lombok.Builder;
 import lombok.Data;
 import lombok.NoArgsConstructor;
 
+import javax.persistence.Column;
 import javax.persistence.Entity;
 import javax.persistence.GeneratedValue;
 import javax.persistence.Id;
@@ -21,6 +22,7 @@ public class Film {
     @GeneratedValue
     private int id;
 
+    @Column(unique = true)
     private String name;
     private String genre;
     private int minAge;
