+++ src/main/java/com/java_school/cinemabot/services/database/DatabaseFilmService.java
@@ -6,8 +6,6 @@ import java.util.List;
 
 public interface DatabaseFilmService {
     List<Film> getAllFilms();
-
     List<Film> getFilmsByGenre(String genre);
-
     void saveFilms();
 }
