+++ src/main/java/com/java_school/cinemabot/services/external/ExternalSessionService.java
@@ -1,5 +1,7 @@
 package com.java_school.cinemabot.services.external;
 
+import com.java_school.cinemabot.services.external.model.ExternalSession;
+
 import java.util.List;
 
 public interface ExternalSessionService {
